Add PAT Provisioning
Okay so on reflection having an API for PATs is just wrong as you'd
eventually have to share a refresh token, which being single use would
log you out as soon as tooling refreshed, or vice versa, or the token
was revoked.  So it's common sense that you do the oauth2 flow again in
order to start a new ancestry of tokens.  Luckily we can just reuse
what's already there and use the client state to control how the
resulting token is handled.