Improve List Views More (#57)
What was simply an addition of the provisioning status and a resource
age turned out to be a bit more involved.  When using the OpenAPI
`--generate-alias-as-model` flag, it generated a type for things like
`Array<string>`, but it also generated ones for
`Array<OrganiationRead>`, the problem with the latter is that when it
unmarshaled the JSON, it doesn't parse the items at all, so where you'd
expect a `Date` to be that, it was just returned as a string, so
typescript and javascript were totally out of whack.  By removing this
flag it forces the generator to correctly parse everything, at the
expense of sligntly less terse code.