Make Flavor Selection "Just Better" (TM)
Presently when selecting a flavor, we see a list of sausages... Okay not
strictly true, we can also see the CPU/RAM/GPU topology too, which
brings us to the point of, with all this metadata, do we (or an end
user) really care what it's called, as apposed to what it can do?  No is
the succinct answer.