Start RBAC Work (#89)
Users have access to APIs based on RBAC, so a non-admin user will not be
able to see group memberships, edit projects etc.  This introduces an
initial framework to block access to these APIs (thus avoiding API error
toasts), and instead present an error screen.  Obviously this is an
initial hack to see what works and what doesn't, so liable to change,
but it appears to do what we want it to!  Additionally fixes a problem
with flavor selection not working due to the IDs not being properly
scoped.