Remove Sensitive Google Scopes (#122)
With implicit groups, that requires a sensitive scope from Google, and
they deem it necessary to make you record a video demoing your app so
they can act as judge, jury and executioner as to whether or not it's
allowed without a bunch of warnings.  Removing this removes that
barrier.