Work Around Race
There is an apparent race with load functions as part of an async
invalidate, and them running as pert of a navigation that causes a bit
of a lock up (especially when you have a div blocking everything!).
This makes it less likely to occur, but I suspect it can still happen if
one's still running when the navigation is triggered.