Invalidate Race Workaround (#138)
What feels like a proper fix for the prior commit.  Based on feedback
from https://github.com/sveltejs/kit/issues/9354 it appears Svelte gives
up if an invalidation happens during a navigation.  Now the user could
reclick the button until it works, but it's not intuitive.  Even worse
for us, we have a full screen div blocking the UI indicating something
is happening!