Allow Client ID/Secret From a Secret (#177)
And it always comes from a secret so you can control who can see it via
Kyverno and friends.  We lose the convenience method of supplying a
client name as you need to supply a secret anyway, so may as well grab
that at the same time.  You can also override secret creation and supply
your own from the External Secrets Operator etc.