Show percentage not value in dashboard (#211)
https://www.skeleton.dev/docs/components/progress-ring/svelte#api-reference
suggests that the ProgressRing will calculate the proportion it
displays inside the ring, given the `value` and `max`; however, what
it _does_ is show the value, with a percentage sign.

To make sure we're showing a percentage, actually calculate it as a
percentage and let `max` default to `100` (and format to
"percentage").